JS-7 Mosaic Restriction - Added Copyright and TransactionMappers, and moved 2 classes to correct location
