+++ sdk-core/src/main/java/io/nem/sdk/model/transaction/MosaicAddressRestrictionTransaction.java
@@ -1,12 +1,25 @@
+/*
+ * Copyright 2019. NEM
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
 package io.nem.sdk.model.transaction;
 
 import io.nem.catapult.builders.AmountDto;
-import io.nem.catapult.builders.BlockDurationDto;
 import io.nem.catapult.builders.KeyDto;
 import io.nem.catapult.builders.MosaicAddressRestrictionTransactionBuilder;
-import io.nem.catapult.builders.MosaicDefinitionTransactionBuilder;
-import io.nem.catapult.builders.MosaicIdDto;
-import io.nem.catapult.builders.MosaicNonceDto;
 import io.nem.catapult.builders.SignatureDto;
 import io.nem.catapult.builders.TimestampDto;
 import io.nem.catapult.builders.UnresolvedAddressDto;
